requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
groq==0.4.1
python-dotenv==1.0.0
csv
os
time
random
re